# src/app/api/openapi.yaml
openapi: 3.0.3
info:
  title: Icarus Game Manager API
  description: |
    API for managing Icarus game servers, mods, and configurations.
    
    This API provides endpoints for:
    - Server management (start/stop/restart)
    - Real-time log monitoring
    - Mod installation and management
    - Server configuration
    - Backup and restore operations
  version: 1.0.0
  contact:
    name: Icarus Game Manager
  license:
    name: MIT
servers:
  - url: /api
    description: Local development server
...
paths:
...
  /backups/settings:
    get:
      tags:
        - Backups
      summary: Get backup settings
      description: Retrieve current backup configuration settings
      responses:
        '200':
          description: Backup settings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  settings:
                    $ref: '#/components/schemas/BackupSettings'
        '500':
          description: Failed to get backup settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Backups
      summary: Update backup settings
      description: Update backup configuration settings including auto backup and max backups limit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupSettings'
      responses:
        '200':
          description: Backup settings updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  settings:
                    $ref: '#/components/schemas/BackupSettings'
        '400':
          description: Invalid settings format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to update backup settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...
    BackupSettings:
      type: object
      properties:
        autoBackup:
          type: boolean
          description: Whether to automatically create backups before server start
        maxBackups:
          type: integer
          minimum: 1
          maximum: 100
          description: Maximum number of backups to keep (older backups will be automatically deleted)
      required:
        - autoBackup
        - maxBackups